Updates for macOS 10.15 (Catalina), various improvements
Use an icon that better matches the visual style of Catalina's toolbar icons.
Use a dark icon when macOS is running in dark mode (at compile time, not runtime,
so a recompile is needed when macOS switches light/dark mode).

Open new windows via the "open" CLI tool, passing a working directory to it;
it's more reliable than asking Terminal to run shell script after its window has opened,
and there's no flash of output while the working directory is changed.
(We still need to ask Terminal to run shell script when opening a new tab, though.)

When opening a tab, if Terminal is already running, use "open -a Terminal" to activate it,
instead of telling Terminal to "activate", so only the frontmost window is brought forward,
letting the rest stay where they were in the z-order, rather than yanking them forward too.

Use a leading space for the "cd" command, to avoid storing it in shell history
(if the shell has been configured to match).

Clear Terminal's scrollback buffer when we clear its screen.